refactor : 디렉토리 변경으로 인한 영향 미친 부분 변경
